Move filter tests into their own file
So that setup and teardown can include all that is common for filter
tests
